async_http_client: Add single shot mechanism.
As closing the TCP connection happens asynchronously, if the API is
used fast enough for a new request, the connection is still open.
It then can happen, that the next request is sent over the old
connection, but its response is lost, as the TCP connection is then
closed.

Fix this by introducing a single shot member that is set if
a request with "Connection: Close" is sent or a response with the same
header is received. The client then blocks creating new requests
until clear is called, thus enforcing creating a new TCP connection.