ringbuffer: Allow using another type as storage.
TF_RingBuffer now can store elements of a smaller type (f.e. char)
in a buffer of a larger type (f.e. uint32_t). The ring buffer ensures
that only reads and writes aligned to the buffer type are done.

This enables us using the 32-bit accessible memory for ring buffers
and thus frees up some heap storage for other purposes.

The event log is modified to store it's contents in the 32-bit
accessible memory.