web_sockets: Check if socket is a web socket before sending.
We then don't have to remove stale work queue items and are sure
that frames are only sent to web socket fds. This fixes the issue
with an fd that was closed and reopened, but only if we assume
that all messages can be broadcasted to all web socket connections.