api: Change naming schema. Complain on too long config paths.
The maximum file name length is 32 characters minus
- the null terminator
- the leading (required!) /
- a marker for temporary files
Thus the old naming schema using .json for configs and .json.tmp
for their temporary files was really wasteful, leaving only 21
characters for the name itself.

Drop the extensions and use a leading dot for temp files. This allows
using up to 29 characters for the config file name itself.

Also add a return type to addPersistentConfig and complain if the
path can't be used.