web_socket: Fix authentication.
We have to check the authentication headers before doing the web
socket handshake. However the handler is only called after the
handshake if it is marked as is_websocket. We remove the marker
and handle the handshake ourselves. As the required methods are
declared in private headers, we copy those over. This makes it easier
to stay in sync with upstream, as we can just overwrite the headers.

Note that this change is not tested yet, as recompiling the
libs is currently broken upstream.